원 그래프, 막대 그래프 데이터 바꾸기(달마다 각 감정의 별점의 평균과 감정 개수 백분율 계산하기)

+++ src/main/java/com/example/emotrak/dto/graph/GraphIdCountResponseDto.java
@@ -9,10 +9,10 @@ import lombok.Setter;
 @NoArgsConstructor
 public class GraphIdCountResponseDto {
     private Long id;
-    private Long count;
+    private float count;
     private float percentage;
 
-    public GraphIdCountResponseDto(Long id, Long count, float percentage){
+    public GraphIdCountResponseDto(Long id, float count, float percentage){
         this.id = id;
         this.count = count;
         this.percentage = percentage;
